swizzle the app delegate methods as well
